OD-1728: Perform steps O, P, Q, R automatically
This is a relatively straight-forward migration of the checks that are
supposed to be manually performed for O, P, Q, and R.

The only significant thing that's worth highlighting is the choice to
use length() instead of char_length() for step R.  In the case of
very large archives, I'm concerned that the chapter length check will
start to be consumed by the unicode functions inside the database.  I'm
proactively choosing this, since I think it's few and far in between for
chapters to be larger than 250,000 characters, nevertheless 250,000
multibyte characters.  Still, if this becomes a frequent false positive,
it may be worth reassessing.

<!-- ps-id: 7701a9dc-3f5d-4272-abb4-8cd6d2b5e61f -->