08: Check for empty/malformed chapters, authors
This check replaces what is currently known as steps G (invalid author
names), I (invalid author email addresses), and J (empty chapters).
Note, it does not confirm that email addresses are in any way valid, or
even properly formatted; merely that the import of whatever is in the
column won't break the archive.